Prevent adding duplicate tags to the same group
This PR prevents users from calling add() into an ObservableGroup
if an Observable has been already added before and is still in flight.
This prevents potential harmful and unwanted behavior we had before
that would cancel any existing previous requests. This is dangerous
and we don't want to silently do that, since it may lead to subtle
bugs.